refactor: migrate lub/crudRepository to ts
